Reorder currencies after selecting new base

+++ app/src/main/java/com/marzec/ui/main/ConverterPresenter.kt
@@ -5,13 +5,14 @@ import com.marzec.extensions.ioTransform
 import com.marzec.model.Rate
 import com.marzec.usecase.CurrencyConverterUseCase
 import io.reactivex.disposables.Disposable
+import java.math.BigDecimal
 import javax.inject.Inject
 
 class ConverterPresenter @Inject constructor(
         private val converterUseCase: CurrencyConverterUseCase
 ) : ConverterContract.Presenter {
 
-    private val defaultRate = Rate("EUR", 1.0)
+    private val defaultRate = Rate("EUR", BigDecimal.valueOf(1))
 
     private var disposable: Disposable? = null
     private var view: ConverterContract.View? = null
